requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
click==8.1.3
dnspython==2.4.2
Flask==2.2.3
Flask-Cors==4.0.0
Flask-JWT==0.3.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.2
PyJWT==2.8.0
pymongo==4.6.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
Werkzeug==2.2.3
